+++ Sparky/src/Platform/OpenGL/OpenGLShader.h
@@ -24,7 +24,7 @@ namespace Sparky {
 		void SetBool(const std::string& name, bool value) const override;
 		void SetInt(const std::string& name, int value) const override;
 		void SetIntArray(const std::string& name, int* data, uint32_t count) const override;
-		void SetFloat(const std::string& name, int value) const override;
+		void SetFloat(const std::string& name, float value) const override;
 		void SetMat3(const std::string& name, const Math::mat3& matrix) const override;
 		void SetMat4(const std::string & name, const Math::mat4 & matrix) const override;
 		void SetFloat2(const std::string & name, const Math::vec2 & vector) const override;
