+++ Sparky/src/Sparky/Renderer/Shader.h
@@ -19,7 +19,7 @@ namespace Sparky {
 		virtual void SetBool(const std::string& name, bool value) const = 0;
 		virtual void SetInt(const std::string& name, int value) const = 0;
 		virtual void SetIntArray(const std::string& name, int* data, uint32_t count) const = 0;
-		virtual void SetFloat(const std::string& name, int value) const = 0;
+		virtual void SetFloat(const std::string& name, float value) const = 0;
 		virtual void SetMat3(const std::string& name, const Math::mat3& matrix) const = 0;
 		virtual void SetMat4(const std::string& name, const Math::mat4& matrix) const = 0;
 		virtual void SetFloat2(const std::string& name, const Math::vec2& vector) const = 0;
